+++ ui/src/routes/indicators/[id]/index.svelte
@@ -108,9 +108,8 @@
 		api_type,
 		auth_provider,
 		data_date,
-		description_short,
-		description_long,
 		description,
+		title,
 		endpoint_url,
 		is_public,
 		query,
@@ -118,6 +117,7 @@
 		schema,
 		source_name,
 		source_url,
+		subtitle,
 		url,
 		warning,
 		year_extent,
@@ -241,14 +241,14 @@
 </script>
 
 <svelte:head>
-	<title>BEIS indicators - {description_short}</title>
+	<title>BEIS indicators - {subtitle}</title>
 </svelte:head>
 
 <div class='container'>
 	<header>
 		<div>
-			<h1>{description_short}</h1>
-			<p>{description}</p>
+			<h1>{subtitle}</h1>
+			<p>{title}</p>
 		</div>
 		<div on:click={toggleInfoModal}>
 			<IconInfo
@@ -451,7 +451,7 @@
 			{api_type}
 			{auth_provider}
 			{data_date}
-			{description_long}
+			{description}
 			{endpoint_url}
 			{is_public}
 			{query}